Fix GlockenSpool Bug in Mount
- fix glockenspool bug
- add favicon
- remove redundant files
