+++ src/components/lab/GlockenSpool.vue
@@ -26,8 +26,13 @@
   </section>
 </template>
 <script>
-import { onMounted, reactive } from '@vue/composition-api'
-import { useAudioGrid } from '@/lib/glocken-spool.setup'
+import { onMounted, ref } from '@vue/composition-api'
+import randomBytes from 'randombytes'
+import random from 'canvas-sketch-util/random'
+import palettes from 'nice-color-palettes'
+import { makeGrid } from '@/lib/canvas.helper'
+import { playNote, scale as scaleData } from '@/lib/sound.helper'
+
 export default {
   name: 'glocken-spool',
   props: {
@@ -41,14 +46,50 @@ export default {
     }
   },
   setup (props) {
-    let vm = reactive({})
+    const points = ref([])
+    const palette = ref(random.pick(palettes))
+    const audio = ref(null)
+    const scale = ref(scaleData)
+    const playNoteFn = ref(() => {})
 
     onMounted(() => {
-      vm = useAudioGrid(props)
+      if (!window) {
+        console.log('no window')
+        return
+      }
+
+      audio.value = new (window.AudioContext || window.webkitAudioContext)()
+      playNoteFn.value = playNote(audio.value)
+      generate()
     })
 
     return {
-      ...vm
+      points,
+      palette,
+      audio,
+      generate,
+      playNote: playNoteFn
+    }
+
+    function makePoints (points, radiusMultiplier) {
+      return makeGrid(props.dimensions)
+        .filter(() => random.value() > 0.5)
+        .map(points => ({
+          ...points,
+          radius: Math.abs(random.gaussian() * props.radiusMultiplier),
+          fill: random.pick(palette.value),
+          freq: random.pick(scale.value),
+          id: randomBytes(8).toString('hex')
+        }))
+    }
+
+    function makePalette () {
+      return random.pick(palettes)
+    }
+
+    function generate () {
+      points.value = makePoints()
+      palette.value = makePalette()
     }
   }
 }
