--- static/README.md
@@ -1,3 +1,0 @@
-Add static files here. Files in this directory will be copied directly to `dist` folder during build. For example, /static/robots.txt will be located at https://yoursite.com/robots.txt.
-
-This file should be deleted.
